+++ core/src/main/java/org/opentele/consult/repository/ClientRepository.java
@@ -16,16 +16,16 @@ public interface ClientRepository extends AbstractRepository<Client> {
             "join organisation o on c.organisation_id = o.id " +
             "where c.name ilike :s and o.id = :organisationId and c.id not in (select client_id from appointment where consultation_room_id = :consultationRoomId ) order by c.name limit 10",
             nativeQuery = true)
-    List<Client> searchClients(String s, int organisationId, int consultationRoomId);
-    default List<Client> getClientsMatching(String s, Organisation organisation, int consultationRoomId) {
+    List<Client> searchClients(String s, Long organisationId, long consultationRoomId);
+    default List<Client> getClientsMatching(String s, Organisation organisation, long consultationRoomId) {
         return searchClients("%" + s + "%", organisation.getId(), consultationRoomId);
     }
     @Query(value = "SELECT count(c.id) FROM client c " +
             "join organisation o on c.organisation_id = o.id " +
             "where c.name ilike :s and o.id = :organisationId and c.id not in (select client_id from appointment where consultation_room_id = :consultationRoomId )",
             nativeQuery = true)
-    int countClients(String s, int organisationId, int consultationRoomId);
-    default int countClientsMatching(String s, Organisation organisation, int consultationRoomId) {
+    int countClients(String s, long organisationId, long consultationRoomId);
+    default int countClientsMatching(String s, Organisation organisation, long consultationRoomId) {
         return countClients("%" + s + "%", organisation.getId(), consultationRoomId);
     }
 
@@ -33,7 +33,7 @@ public interface ClientRepository extends AbstractRepository<Client> {
             "join organisation o on c.organisation_id = o.id " +
             "where (c.name ilike :q or COALESCE(c.registration_number, '') ilike :q) and o.id = :organisationId order by c.name limit 10",
             nativeQuery = true)
-    List<Client> searchClients(String q, int organisationId);
+    List<Client> searchClients(String q, long organisationId);
     default List<Client> searchClients(String s, Organisation organisation) {
         return searchClients("%" + s + "%", organisation.getId());
     }
@@ -43,13 +43,13 @@ public interface ClientRepository extends AbstractRepository<Client> {
             "where a.consultation_room_id = :consultationRoomId" +
             " order by c.name",
             nativeQuery = true)
-    List<ConsultationRoomClientProjection> getClients(int consultationRoomId);
+    List<ConsultationRoomClientProjection> getClients(long consultationRoomId);
 
     @Query(value = "SELECT c.* FROM client c " +
             "join organisation o on c.organisation_id = o.id " +
             "where (c.name ilike :name and COALESCE(c.registration_number, '') ilike :registrationNumber) and o.id = :organisationId order by c.name limit 10",
             nativeQuery = true)
-    List<Client> searchClients(String name, String registrationNumber, int organisationId);
+    List<Client> searchClients(String name, String registrationNumber, long organisationId);
     default List<Client> getClientsMatching(String name, String registrationNumber, Organisation organisation) {
         String s2 = registrationNumber == null ? "" : registrationNumber;
         return searchClients("%" + name + "%", "%" + s2 + "%", organisation.getId());
