+++ core/src/main/java/org/opentele/consult/repository/OrganisationRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganisationRepository extends CrudRepository<Organisation, Integer> {
+public interface OrganisationRepository extends CrudRepository<Organisation, Long> {
     Organisation findByName(String name);
-    default Organisation findEntity(int id) {
+    default Organisation findEntity(long id) {
         if (id == 0) return null;
         return findById(id).get();
     }
