+++ server/src/main/java/org/opentele/consult/framework/UserSession.java
@@ -1,6 +1,5 @@
 package org.opentele.consult.framework;
 
-import org.opentele.consult.domain.security.User;
 import org.springframework.context.annotation.Scope;
 import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Component;
@@ -10,22 +9,22 @@ import java.io.Serializable;
 @Component
 @Scope(value="session", proxyMode= ScopedProxyMode.TARGET_CLASS)
 public class UserSession implements Serializable {
-    private int currentOrganisationId;
-    private int userId;
+    private long currentOrganisationId;
+    private long userId;
 
-    public int getCurrentOrganisationId() {
+    public long getCurrentOrganisationId() {
         return currentOrganisationId;
     }
 
-    public void setCurrentOrganisationId(int currentOrganisationId) {
+    public void setCurrentOrganisationId(long currentOrganisationId) {
         this.currentOrganisationId = currentOrganisationId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setCurrentUserId(int userId) {
+    public void setCurrentUserId(long userId) {
         this.userId = userId;
     }
 }
